Gettings things ready to create vertex & index buffer cpu side
