+++ examples/main.rs
@@ -192,6 +192,8 @@ fn main() {
     )
     .unwrap();
 
+    
+
     /*
 
         // load resources with transfer queue
