+++ examples/shaders/shader.vert
@@ -2,9 +2,7 @@
 #extension GL_ARB_separate_shader_objects : enable
 
 layout(location = 0) in vec3 inPosition;
-layout(location = 1) in vec3 inNormal;
-layout(location = 2) in vec4 inColor;
-layout(location = 3) in vec2 inTexCoord;
+layout(location = 1) in vec4 inColor;
 
 layout(location = 0) out vec4 fragColor;
